added Collapsible to description & refactored Description - made it more modular
